Pointer gestures are timed in real time, so they keep working while the game is paused
Long press, the double-click window, swipe duration, both raycasters' hold-to-drag and navigation
repeat were measured on the game clock, which stands still while the game is paused: in a pause menu
a held press never matured, any two clicks on one widget were a double click however slowly they
came, and a held direction never repeated.

Every stamp and every measurement is now on UDreamEventSystem::GetPointerClockSeconds, the world's
real time, as Slate times its own double clicks and key repeat; a render-target surface stamps the
pointer it drives on the same clock. The standalone module keeps its press and release stamps as
doubles, like the event data fields they are copied into. The idle-navigation test advances the real
clock beside the game clock it already advanced.

+++ Source/DreamGUI/Public/Event/InputModule/DreamStandaloneInputModule.h
@@ -69,8 +69,13 @@ protected:
 	struct StandaloneInputData
 	{
 		bool bTriggerPress = false;
-		float PressTime = 0;
-		float ReleaseTime = 0;
+		/**
+		 * Pointer-clock stamps (UDreamEventSystem::GetPointerClockSeconds), as double as the event data
+		 * fields they are copied into: real time only ever grows during a session, and a float of it
+		 * loses milliseconds within the first few hours.
+		 */
+		double PressTime = 0;
+		double ReleaseTime = 0;
 		int PointerID = 0;
 		EDreamUIMouseButtonType MouseButtonType = EDreamUIMouseButtonType::Left;
 		FVector PointerPosition = FVector::ZeroVector;
